+++ src/helpers/style-dictionary/formats/tokens.js
@@ -94,37 +94,6 @@ const customOpacityObjectFormatter = (dictionary, theme, isJS) => {
   );
 };
 
-const colorDocumentationFormatter = (dictionary, isJS) => {
-  const renderDescription = (desc) => {
-    if (desc) {
-      return isJS ? `description: "${desc}"` : `description: string`;
-    }
-    return "";
-  };
-
-  return Object.entries(dictionary.properties.colors)
-    .map((tokens) => {
-      const colorObj = tokens[0];
-      const filteredTokens = dictionary.allTokens.filter(
-        (token) => token.attributes.type === colorObj
-      );
-
-      return (
-        declaration(isJS) +
-        `${colorObj} : {` +
-        filteredTokens.map((token) => {
-          return `${token.name}:{
-            name: ${isJS ? `"${colorObj}.${token.name}"` : `string`},
-            hex: ${valueOrType(token, isJS)},
-            ${renderDescription(token.description)}
-          }`;
-        }) +
-        `}${commaOrColon(isJS)}`
-      );
-    })
-    .join(`\n`);
-};
-
 StyleDictionary.registerFormat({
   name: "custom/format/typescript-color-declarations",
   formatter: ({ dictionary, file }) => {
@@ -148,24 +117,5 @@ StyleDictionary.registerFormat({
       customOpacityObjectFormatter(dictionary, file, true) +
       `};`
     );
-  },
-});
-
-StyleDictionary.registerFormat({
-  name: "custom/format/javascript-colors-documentation",
-  formatter: ({ dictionary, file }) => {
-    return (
-      fileHeader({ file }) +
-      `module.exports = {` +
-      colorDocumentationFormatter(dictionary, true) +
-      `};`
-    );
-  },
-});
-
-StyleDictionary.registerFormat({
-  name: "custom/format/typescript-color-declarations-documentation",
-  formatter: ({ dictionary, file }) => {
-    return fileHeader({ file }) + colorDocumentationFormatter(dictionary, false);
   },
 });
