+++ src/helpers/style-dictionary/index.js
@@ -1,0 +1,5 @@
+require("./filters/filters");
+require("./transforms/transforms");
+require("./transforms/transform-groups");
+require("./formats/tokens");
+require("./formats/documentation");
