+++ src/helpers/style-dictionary/transforms/transform-groups.js
@@ -1,0 +1,46 @@
+/**
+ * used in build.js
+ *
+ * transforms are performed sequentially (see transform groups below).
+ * order of operations: filters > transforms > formats
+ *
+ * https://amzn.github.io/style-dictionary/#/transforms?id=transforms
+ */
+
+require("./transforms");
+
+const StyleDictionary = require("style-dictionary");
+
+StyleDictionary.registerTransformGroup({
+  name: "custom/aviary",
+  transforms: [
+    "attribute/cti",
+    "custom/name/remove-desktop-prefix",
+    "custom/name/remove-color-prefix",
+  ],
+});
+
+StyleDictionary.registerTransformGroup({
+  name: "custom/documentation",
+  transforms: [
+    "attribute/cti",
+    "custom/name/remove-desktop-prefix",
+    "custom/name/remove-color-prefix",
+  ],
+});
+
+StyleDictionary.registerTransformGroup({
+  name: "custom/native",
+  transforms: [
+    "attribute/cti",
+    "custom/name/remove-desktop-prefix",
+    "custom/value/rm-px",
+    "custom/value/font-weight-to-string",
+    "custom/name/remove-color-prefix",
+  ],
+});
+
+StyleDictionary.registerTransformGroup({
+  name: "custom/scss",
+  transforms: ["attribute/cti", "custom/name/remove-desktop-prefix"],
+});
